% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{build_df_timer_day}
\alias{build_df_timer_day}
\title{Build the df_timer for days}
\usage{
build_df_timer_day(df_timer, cfg)
}
\arguments{
\item{df_timer}{df_timer}

\item{cfg}{config list}
}
\value{
data.frame
}
\description{
Build the df_timer for days
}
