% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{build_df_timer_intv}
\alias{build_df_timer_intv}
\title{Build the df_timer of interview}
\usage{
build_df_timer_intv(df_timer_ssn, df_timer_day, cfg)
}
\arguments{
\item{df_timer_ssn}{df_timer_ssn}

\item{df_timer_day}{df_timer_day}

\item{cfg}{config list}
}
\value{
data.frame
}
\description{
Build the df_timer of interview
}
