% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{build_df_timer_ssn}
\alias{build_df_timer_ssn}
\title{Build the df_timer of sessions}
\usage{
build_df_timer_ssn(df_timer, df_timer_sctn, cfg)
}
\arguments{
\item{df_timer}{df_timer}

\item{df_timer_sctn}{df_timer_sctn}

\item{cfg}{config list}
}
\value{
data.frame
}
\description{
Build the df_timer of sessions
}
