% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{classify_variables}
\alias{classify_variables}
\title{Classify all variable of a data.frame}
\usage{
classify_variables(df, threhold = 15)
}
\arguments{
\item{df}{A data frame}
}
\value{
a data frame
}
\description{
Classify all variable of a data.frame
}
\examples{
classify_variables(iris)
}
