% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_vars}
\alias{combine_vars}
\title{Combine several variables in one key (multi-level wave or filter)}
\usage{
combine_vars(df, vars, sep = " / ")
}
\arguments{
\item{df}{data.frame}

\item{vars}{vector of the variable of each level}

\item{sep}{separator between the levels}
}
\value{
vector of the complete keys
}
\description{
Used when var_wave (or var_filter) contains more than one variable in the
configuration file, for example the year and the quarter. The complete key
of a row is then "2024 / T1".
}
\examples{
head(combine_vars(mtcars,c("cyl","gear")))
}
