% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{compute_timer_groups}
\alias{compute_timer_groups}
\title{Compute timer of group}
\usage{
compute_timer_groups(df_timer, cfg)
}
\arguments{
\item{df_timer}{df_timer}

\item{cfg}{config list}
}
\value{
data.frame
}
\description{
Compute timer of group
}
