% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{correct_list_df}
\alias{correct_list_df}
\title{corrections of each df of a list}
\usage{
correct_list_df(list_df)
}
\arguments{
\item{list_df}{list of df}
}
\value{
list
}
\description{
corrections of each df of a list
}
