% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_eusilc_sim}
\alias{create_eusilc_sim}
\title{Simulate EU-SILC dataset with injected errors}
\usage{
create_eusilc_sim()
}
\value{
data.frame from laeken::eusilc with ids and errors
}
\description{
Simulate EU-SILC dataset with injected errors
}
\examples{
# create_eusilc_sim()
}
