% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_safe}
\alias{cut_safe}
\title{Cut a continuous variable in n categories.}
\usage{
cut_safe(x, n_breaks = 5)
}
\arguments{
\item{x}{vector to cut}

\item{n_breaks}{number of breaks}
}
\value{
vector
}
\description{
Cut a continuous variable in n categories.
}
