% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_variable_dt}
\alias{describe_variable_dt}
\title{Description d'une variable rendue en DT}
\usage{
describe_variable_dt(x, by = NULL, ..., caption = NULL)
}
\arguments{
\item{x}{vecteur à décrire}

\item{by}{(optionnel) vecteur de même longueur, un bloc de colonnes par valeur}

\item{caption}{(optionnel) titre du tableau}
}
\value{
datatable
}
\description{
Description d'une variable rendue en DT
}
