% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_as_na}
\alias{empty_as_na}
\title{Replace empty by na}
\usage{
empty_as_na(vec)
}
\arguments{
\item{vec}{Vector of values}
}
\value{
Vector
}
\description{
Replace empty by na
}
\examples{
airquality[which(is.na(airquality$Ozone)), "Ozone"] <- ""
empty_as_na(airquality$Ozone)
}
