% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{extract_config}
\alias{extract_config}
\title{extract a config from key (config from load_config)}
\usage{
extract_config(config, key_)
}
\arguments{
\item{config}{df of configuration}

\item{key_}{key to extract}
}
\value{
string
}
\description{
extract a config from key (config from load_config)
}
