% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{find_timer_rds}
\alias{find_timer_rds}
\title{Extract timer rds files from folder}
\usage{
find_timer_rds(folder)
}
\arguments{
\item{folder}{folder where check the file}
}
\value{
path of files
}
\description{
Extract timer rds files from folder
}
