% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{folder_to_df}
\alias{folder_to_df}
\title{tranform data from folder to config and df}
\usage{
folder_to_df(folder, file_pattern = "*.csv", file_config = "config.txt")
}
\arguments{
\item{folder}{folder of databases}

\item{file_pattern}{pattern of the databases (*.csv by default)}

\item{file_config}{name of the configuration file (config.txt by default)}
}
\value{
list(df,configs)
}
\description{
tranform data from folder to config and df
}
\examples{
\dontrun{
folder_to_df("ESS10")
}
}
