% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{group_vs_others}
\alias{group_vs_others}
\title{Vecteur de groupe "VAR = valeur" / "VAR != valeur" pour les comparaisons}
\usage{
group_vs_others(x, value, label)
}
\arguments{
\item{x}{Vecteur a regrouper}

\item{value}{valeur discrétionnaire}

\item{label}{label à utiliser}
}
\value{
vector
}
\description{
Vecteur de groupe "VAR = valeur" / "VAR != valeur" pour les comparaisons
}
