% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{heat_map_group}
\alias{heat_map_group}
\title{Create a heatmap}
\usage{
heat_map_group(db_stat, threshold)
}
\arguments{
\item{db_stat}{data frame from preparation_stats function}

\item{threshold}{threshold to show difference}
}
\value{
heatmap (ggplot)
}
\description{
Create a heatmap
}
\examples{
library(laeken)
data(eusilc)

info_vars <- classify_variables(eusilc)
vars_vd <- info_vars[info_vars$type == "Modal", ]$variable
vars_vc <- info_vars[info_vars$type == "Continuous", ]$variable
db_stats <- prepa_stats(eusilc, vars_vd, vars_vc, "db040")
heat_map_group(db_stats, 5)
}
