% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_itw.R
\name{index_homogeneity}
\alias{index_homogeneity}
\title{index_homogeneity}
\usage{
index_homogeneity(df, var_itw, threahold = 30)
}
\arguments{
\item{df}{dataset}

\item{var_itw}{groupe variable}

\item{threahold}{threahold}
}
\value{
dataset
}
\description{
index_homogeneity
}
