% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.integer64}
\alias{is.integer64}
\title{Check if value is integer64}
\usage{
is.integer64(x)
}
\arguments{
\item{x}{Value}
}
\value{
Boolean
}
\description{
Check if value is integer64
}
\examples{
is.integer64(c(1:100)) # FALSE
}
