% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{key_level1}
\alias{key_level1}
\title{First level of a multi-level key}
\usage{
key_level1(x, sep = " / ")
}
\arguments{
\item{x}{vector of keys}

\item{sep}{separator between the levels}
}
\value{
vector of the first level values
}
\description{
First level of a multi-level key
}
\examples{
key_level1(c("2024 / T1","2024 / T2","2023 / T4"))
}
