% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{key_level2}
\alias{key_level2}
\title{Other levels of a multi-level key (used as label in the interface)}
\usage{
key_level2(x, sep = " / ")
}
\arguments{
\item{x}{vector of keys}

\item{sep}{separator between the levels}
}
\value{
vector of the values after the first level
}
\description{
Other levels of a multi-level key (used as label in the interface)
}
\examples{
key_level2(c("2024 / T1","2024 / T2","2023 / T4"))
}
