% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keys_vars}
\alias{keys_vars}
\title{Keys (modalities) of a wave or filter variable}
\usage{
keys_vars(df, vars, var_key, level = "all", sep = " / ")
}
\arguments{
\item{df}{data.frame}

\item{vars}{variables of each level (see vars_levels)}

\item{var_key}{name of the variable containing the complete key}

\item{level}{1 (first level only), 2 (complete key) or "all" (both)}

\item{sep}{separator between the levels}
}
\value{
vector of keys
}
\description{
Keys (modalities) of a wave or filter variable
}
\examples{
keys_vars(mtcars,"cyl","cyl")
}
