% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_dist}
\alias{list_dist}
\title{List distribution of discrete variables}
\usage{
list_dist(df, vars_vd)
}
\arguments{
\item{df}{data.frame}

\item{vars_vd}{vector of discrete variables}
}
\value{
list
}
\description{
List distribution of discrete variables
}
\examples{
list_dist(mtcars,c("cyl","vs","gear"))
}
