% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa_dt.R
\name{list_dist_dt}
\alias{list_dist_dt}
\title{List of proportions for categorical variables}
\usage{
list_dist_dt(df, vars_discretes, na.rm = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{vars_discretes}{vector of categorical variable to calculate proportion}

\item{na.rm}{Remove missing values or not ?}
}
\value{
list
}
\description{
List of proportions for categorical variables
}
