% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{load_config}
\alias{load_config}
\title{load a config file for prepare data}
\usage{
load_config(file_path)
}
\arguments{
\item{file_path}{path of the configuration file}
}
\value{
df
}
\description{
load a config file for prepare data
}
