% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_keys}
\alias{match_keys}
\title{Rows of a data.frame matching one or several keys}
\usage{
match_keys(df, vars, var_key, keys, sep = " / ")
}
\arguments{
\item{df}{data.frame}

\item{vars}{variables of each level (see vars_levels)}

\item{var_key}{name of the variable containing the complete key}

\item{keys}{vector of keys to keep}

\item{sep}{separator between the levels}
}
\value{
logical vector
}
\description{
A row matches a key when its complete key is one of the keys, or when its
first level is one of the keys (selection of a year without the quarter).
}
\examples{
match_keys(mtcars,"cyl","cyl",4)
}
