% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{prepa_all_surveys}
\alias{prepa_all_surveys}
\title{Preparation of all surveys from a folder}
\usage{
prepa_all_surveys(folder)
}
\arguments{
\item{folder}{folder of the folders of survey}
}
\value{
NULL (creation of rds)
}
\description{
Preparation of all surveys from a folder
}
\examples{
\dontrun{
prepa_all_surveys("shiny-examples/complete/data")
}
}
