% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{prepa_stats}
\alias{prepa_stats}
\title{Create a summarise of all the difference}
\usage{
prepa_stats(db, vars_vd, vars_vc, var_group)
}
\arguments{
\item{db}{data frame for the summary}

\item{vars_vd}{Vector of discrete variables}

\item{vars_vc}{Vector of continuous variables}

\item{var_group}{Name of group variable}
}
\value{
data frame
}
\description{
Create a summarise of all the difference
}
\examples{
library(laeken)
data(eusilc)

info_vars <- classify_variables(eusilc)
vars_vd <- info_vars[info_vars$type == "Modal", ]$variable
vars_vc <- info_vars[info_vars$type == "Continuous", ]$variable
prepa_stats(eusilc, vars_vd, vars_vc, "db040")
}
