% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa_dt.R
\name{prepa_stats_dt}
\alias{prepa_stats_dt}
\title{prepa_stats_dt}
\usage{
prepa_stats_dt(df, var_group, configs, na.rm = FALSE)
}
\arguments{
\item{df}{}

\item{var_group}{}

\item{configs}{}

\item{na.rm}{}
}
\value{
data.frame
}
\description{
prepa_stats_dt
}
