% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{prepa_survey}
\alias{prepa_survey}
\title{Preparation of a survey}
\usage{
prepa_survey(
  folder_path,
  file_pattern = "*.csv",
  file_config = "config.txt",
  na.rm = FALSE
)
}
\arguments{
\item{folder_path}{folder of survey}

\item{file_pattern}{pattern of the databases (*.csv by default)}

\item{file_config}{name of the configuration file (config.txt by default)}
}
\value{
NULL (creation of rds)
}
\description{
Preparation of a survey
}
\examples{
\dontrun{
prepa_survey("shiny/complete/ESS10")
}
}
