% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{prepa_surveys}
\alias{prepa_surveys}
\title{Preparation of all surveys from a folder}
\usage{
prepa_surveys(folder_path, depth_folder = 1, ...)
}
\arguments{
\item{folder_path}{folder of the folders of survey}

\item{depth_folder}{level of depth for the tree structure}

\item{...}{argument to pass to folder_to_df (example : file_pattern)}
}
\value{
NULL (creation of rds)
}
\description{
Preparation of all surveys from a folder
}
\examples{
\dontrun{
prepa_surveys("inst/extdata/SILC/HFILE")
}
}
