% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{prepa_timer_from_df}
\alias{prepa_timer_from_df}
\title{Prepation of timers from raw}
\usage{
prepa_timer_from_df(df_timer_raw, cfg)
}
\arguments{
\item{df_timer_raw}{df_timer_raw}

\item{cfg}{config list}
}
\value{
data.frame
}
\description{
Prepation of timers from raw
}
