% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{read_timer_csv}
\alias{read_timer_csv}
\title{Read a timer file in csv}
\usage{
read_timer_csv(file_timer)
}
\arguments{
\item{file_timer}{path of file}
}
\value{
data.frame
}
\description{
Read a timer file in csv
}
