% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{read_timer_csvs}
\alias{read_timer_csvs}
\title{Read and rbind many timer files in csv}
\usage{
read_timer_csvs(files_timer)
}
\arguments{
\item{files_timer}{path of files}
}
\value{
data.frame
}
\description{
Read and rbind many timer files in csv
}
