% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey}
\alias{runVizsurvey}
\title{Shiny vizsurvey}
\usage{
runVizsurvey(link_data_folder = NULL, is_double_folder = T)
}
\arguments{
\item{link_data_folder}{link to directory of data}

\item{is_double_folder}{is the directory contains subdirectories of survey}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey
}
\examples{
# runVizsurvey()
}
