% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_IQR}
\alias{scale_IQR}
\title{Robust Scale of a varible with IQR}
\usage{
scale_IQR(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Robust Scale of a varible with IQR
}
\examples{
head(scale_IQR(iris$Sepal.Length))
}
