% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{score_isoforest}
\alias{score_isoforest}
\title{calculate isoforest score from df}
\usage{
score_isoforest(df)
}
\arguments{
\item{df}{database}
}
\value{
vector
}
\description{
calculate isoforest score from df
}
\examples{
score_isoforest(iris[sapply(iris, is.numeric)])
}
