% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vars_levels}
\alias{vars_levels}
\title{Variables of each level of the wave or of the filter}
\usage{
vars_levels(configs, type = "wave")
}
\arguments{
\item{configs}{list of configuration}

\item{type}{"wave" or "filter"}
}
\value{
vector of the variables of each level
}
\description{
Kept compatible with the global.rds prepared before the multi-level option :
vars_wave (or vars_filter) is then missing and var_wave is used.
}
\examples{
vars_levels(list(var_wave = "YEAR"),"wave")
}
