% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer2.R
\name{build_df_timer_intvwr}
\alias{build_df_timer_intvwr}
\title{Build the df_timer of interviewer}
\usage{
build_df_timer_intvwr(df_timer_intv, cfg)
}
\arguments{
\item{df_timer_intv}{df_timer_intv}

\item{cfg}{config list}
}
\value{
data.frame
}
\description{
Build the df_timer of interviewer
}
