% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{create_config}
\alias{create_config}
\title{Create a template of configuration file}
\usage{
create_config(
  folder_path,
  file_name = "config.txt",
  name_survey = NULL,
  vars_discretes = NULL,
  vars_continous = NULL,
  var_wave = NULL,
  var_zone = NULL,
  var_group = NULL
)
}
\arguments{
\item{folder_path}{folder where create the file}

\item{file_name}{Name of the config file (config.txt by default)}

\item{name_survey}{Name of the survey (not used)}

\item{vars_discretes}{(optional) preset discretes variables name (VAR1,VAR2,...)}

\item{vars_continous}{(optional) preset continous variables name (VAR1,VAR2,...)}

\item{var_wave}{(optional) variable name of wave}

\item{var_zone}{(optional) variable name of zone}

\item{var_group}{variable name of group}
}
\description{
Create a template of configuration file
}
\examples{
create_config(".") # creation of config.txt in working directory
}
