% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{create_df_stats}
\alias{create_df_stats}
\title{Create statistics from database}
\usage{
create_df_stats(df_, configs, var_calculs, zone_filter = NULL)
}
\arguments{
\item{df_}{database}

\item{configs}{configs}

\item{var_calculs}{variable to create stats}

\item{zone_filter}{(optional) zone modality to filter data}
}
\value{
df
}
\description{
Create statistics from database
}
