% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_fake_silc}
\alias{create_fake_silc}
\title{Use PUF files for SILC example}
\usage{
create_fake_silc(
  vec_country = c("BE", "RO"),
  path_out = "inst/shiny-examples/complete/data/SILC/"
)
}
\arguments{
\item{vec_country}{Vector of country to import}

\item{path_out}{Path to export csv}
}
\value{
nothing, but creation of csv datasets
}
\description{
Use PUF files for SILC example
}
\examples{
# create_fake_silc()
}
