% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{heatmap_group}
\alias{heatmap_group}
\title{Create a heatmap}
\usage{
heatmap_group(df_stats, threshold = 5, color = "red2")
}
\arguments{
\item{df_stats}{data frame from prepa_stats function}

\item{threshold}{threshold to show difference}

\item{color}{color of the cells}
}
\value{
heatmap (ggplot)
}
\description{
Create a heatmap
}
\examples{
library(laeken)
data(eusilc)

df_stats <- prepa_stats(eusilc, "db040")
heatmap_group(df_stats, 5)
}
