% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{loop_stats}
\alias{loop_stats}
\title{Loop of stats creation by filter}
\usage{
loop_stats(df, configs, var_calculs)
}
\arguments{
\item{df}{database}

\item{configs}{configs}

\item{var_calculs}{variable to create stats}
}
\value{
df
}
\description{
Loop of stats creation by filter
}
