% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_chisq_test}
\alias{my_chisq_test}
\title{Specific chisq test to NA and Other modality}
\usage{
my_chisq_test(x, varname, ldist)
}
\arguments{
\item{x}{value to procede chisq test}

\item{varname}{name of the variable}

\item{ldist}{named list of expected probability}
}
\value{
chisq value
}
\description{
Specific chisq test to NA and Other modality
}
\examples{
ldist <- list_dist(mtcars,c("cyl","gear"))
sub_mtcars <- subset(mtcars,vs == 1)
my_chisq_test(sub_mtcars$cyl,"cyl",ldist)
}
