% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{prepa_stats}
\alias{prepa_stats}
\title{Create a summarise of all the difference}
\usage{
prepa_stats(df, var_group, vars_vd = NULL, vars_vc = NULL)
}
\arguments{
\item{df}{data frame for the summary}

\item{var_group}{Name of group variable}

\item{vars_vd}{(optional) Vector of discrete variables}

\item{vars_vc}{(optional) Vector of continuous variables}
}
\value{
data frame
}
\description{
Create a summarise of all the difference
}
\examples{
library(laeken)
data(eusilc)

info_vars <- classify_df(eusilc)
vars_vd <- info_vars[info_vars$type == "Modal", ]$variable
vars_vc <- info_vars[info_vars$type == "Continuous", ]$variable
prepa_stats(eusilc, "db040", vars_vd, vars_vc)
}
