% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepa.R
\name{prepa_survey}
\alias{prepa_survey}
\title{Preparation of a survey}
\usage{
prepa_survey(folder, name_file = NULL, ...)
}
\arguments{
\item{folder}{folder of survey}

\item{name_file}{(optional) name to save file (in the same folder)}

\item{...}{argument to pass to folder_to_df (example : file_pattern)}
}
\value{
NULL (creation of rds)
}
\description{
Preparation of a survey
}
\examples{
\dontrun{
prepa_survey("shiny-examples/complete/ESS10")
}
}
