% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey_from_file}
\alias{runVizsurvey_from_file}
\title{Shiny vizsurvey from a csv/tsv}
\usage{
runVizsurvey_from_file(
  path,
  vars_discretes = NULL,
  vars_continous = NULL,
  var_wave = NULL,
  var_zone = NULL,
  var_group = NULL
)
}
\arguments{
\item{path}{path of a data.frame (can be readed by fread)}

\item{vars_discretes}{(optional) preset of discretes variables}

\item{vars_continous}{(optional) preset of continous variables}

\item{var_wave}{(optional) name of wave variable}

\item{var_zone}{(optional) name of zone variable}

\item{var_group}{(optional) name of group variable}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey from a csv/tsv
}
\examples{
path <- "inst/extdata/SILC/HFILE/BE_2012h_EUSILC.csv"
\dontrun{runVizsurvey_from_file(path,var_group = "NR_ITW",var_zone = "db040")}
}
