% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey_from_file}
\alias{runVizsurvey_from_file}
\title{Shiny vizsurvey from a csv/tsv}
\usage{
runVizsurvey_from_file(
  path,
  var_itw,
  vars_discretes = NULL,
  vars_continous = NULL,
  var_domain = NULL,
  var_group = NULL
)
}
\arguments{
\item{path}{path of a data.frame (can be readed by fread)}

\item{var_itw}{name of interviewer variable}

\item{vars_discretes}{(optional) preset of discretes variables}

\item{vars_continous}{(optional) preset of continous variables}

\item{var_domain}{(optional) name of domain variable}

\item{var_group}{(optional) name of group variable}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey from a csv/tsv
}
\examples{
path <- "inst/shiny-examples/complete/data/ESS/ESS9/ESS9.csv"
\dontrun{runVizsurvey_from_file(path,var_itw = "INTNUM1",var_domain="CNTRY")}
}
