% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey_from_folder}
\alias{runVizsurvey_from_folder}
\title{Shiny vizsurvey with already prepared data}
\usage{
runVizsurvey_from_folder(
  path,
  path_dict = NULL,
  path_nomen = NULL,
  path_archive = NULL,
  data_rds_pattern = "global",
  depth_folder = 1
)
}
\arguments{
\item{data_rds_pattern}{name of the rds file contains all the data}

\item{depth_folder}{level of depth for the tree structure}

\item{link}{link to directory of data}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey with already prepared data
}
\examples{
# We assume that config.txt, and prepa_surveys are already done here.
\dontrun{runVizsurvey_from_folder("inst/extdata",depth_folder = 3)}
}
