% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey_from_folder}
\alias{runVizsurvey_from_folder}
\title{Shiny vizsurvey with already prepared data}
\usage{
runVizsurvey_from_folder(
  link,
  data_rds_pattern = "global",
  is_double_folder = T
)
}
\arguments{
\item{link}{link to directory of data}

\item{data_rds_pattern}{name of the rds file contains all the data}

\item{is_double_folder}{is the directory contains subdirectories of survey}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey with already prepared data
}
\examples{
# We assume that config.txt, and prepa_surveys are already done here.
\dontrun{runVizsurvey_from_folder("inst/extdata",is_double_folder = T)}
}
