% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey_from_r}
\alias{runVizsurvey_from_r}
\title{Shiny vizsurvey from a R data.frame}
\usage{
runVizsurvey_from_r(
  df,
  vars_discretes = NULL,
  vars_continuous = NULL,
  var_wave = NULL,
  var_filter = NULL,
  var_intvwr = NULL
)
}
\arguments{
\item{df}{data.frame}

\item{vars_discretes}{(optional) preset of discretes variables}

\item{var_wave}{(optional) name of wave variable}

\item{var_filter}{(optional) name of filter variable}

\item{var_intvwr}{(optional) name of interviewer variable}

\item{vars_continous}{(optional) preset of continous variables}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey from a R data.frame
}
\examples{
library(laeken)
data(eusilc)
set.seed(123)
eusilc$NR_ITW <- paste(eusilc$db040,sample(1:5,nrow(eusilc),replace = TRUE),sep="-")
\dontrun{runVizsurvey_from_r(eusilc,var_intvwr = "NR_ITW",var_filter = "db040")}
}
