% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appComplete.R
\name{runVizsurvey_from_r}
\alias{runVizsurvey_from_r}
\title{Shiny vizsurvey from a R data.frame}
\usage{
runVizsurvey_from_r(
  df,
  vars_discretes = NULL,
  vars_continous = NULL,
  var_wave = NULL,
  var_zone = NULL,
  var_group = NULL
)
}
\arguments{
\item{df}{data.frame}

\item{vars_discretes}{(optional) preset of discretes variables}

\item{vars_continous}{(optional) preset of continous variables}

\item{var_wave}{(optional) name of wave variable}

\item{var_zone}{(optional) name of zone variable}

\item{var_group}{(optional) name of group variable}
}
\value{
shinyapp
}
\description{
Shiny vizsurvey from a R data.frame
}
\examples{
library(laeken)
data(eusilc)
set.seed(123)
eusilc$NR_ITW <- paste(eusilc$db040,sample(1:5,nrow(eusilc),replace = TRUE),sep="-")
\dontrun{runVizsurvey_from_r(eusilc,var_group = "NR_ITW",var_zone = "db040")}
}
