% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsurvey-package.R
\docType{package}
\name{vizsurvey-package}
\alias{vizsurvey}
\alias{vizsurvey-package}
\title{vizsurvey: Visualisation And Analysis During a Survey Field}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Thomas Delclite \email{thomas.delclite@economie.fgov.be}

Authors:
\itemize{
  \item Adrien Mierop \email{adrien.mierop@economie.fgov.be}
}

}
\keyword{internal}
