% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsurvey-package.R
\docType{package}
\name{vizsurvey-package}
\alias{vizsurvey}
\alias{vizsurvey-package}
\title{vizsurvey: Visualisation And Analysis During a Survey Field}
\description{
vizsurvey is an R package designed to streamline the quality assessment of survey data by providing intuitive visual diagnostics through an interactive dashboard. vizsurvey is especially useful for institutions or researchers conducting large-scale surveys with multiple interviewers, enabling a fast and systematic overview of data quality over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdelc/vizsurvey}
  \item \url{https://tdelc.github.io/vizsurvey/}
  \item Report bugs at \url{https://github.com/tdelc/vizsurvey/issues}
}

}
\author{
\strong{Maintainer}: Thomas Delclite \email{thomas.delclite@economie.fgov.be}

Authors:
\itemize{
  \item Adrien Mierop \email{adrien.mierop@economie.fgov.be}
}

}
\keyword{internal}
